+++ zhmcclient/__init__.py
@@ -56,3 +56,4 @@ from ._storage_volume import *         # noqa: F401
 from ._virtual_storage_resource import *        # noqa: F401
 from ._storage_group_template import *          # noqa: F401
 from ._storage_volume_template import *         # noqa: F401
+from ._capacity_group import *         # noqa: F401